refactor BarChart
With all the charting features we've added, this file has gotten fairly long and messy.
To keep things tidy, let's extract some of these functions to a separate file called charting.ts.

In the process, let's ensure the BarChart is properly typed. Upon reading documentation, I found out that we can type the chart data by using ChartData<'bar', {Z}> where Z is the values that each data point has (in our case, x and y values, which can be number | string)

This makes all the type definitions line up - before this, we had a lot of different type warnings because TypeScript didn't understand our data structure.
